chore(test): Add .defined matcher
Some tests (spec/features/modeling/DeleteElementsSpec.js is one example)
expects some value .to.be.defined or .to.not.be.defined.
But Chai does not have a .defined assertion. It's called .undefined.

Since .defined is undefined, the assertion is a no-op and won't detect
broken code.

I have tested this by commenting out the WHEN statement in
DeleteElementsSpec.js. Without this fix, the test will then silently
continue. With this fix the error is detected as expected.