fix(CommandStack): Given array, dirty elements should exclude undefined
This commit will prevent undefined elements to being concated to
CommandStack's dirty collection.

If the command handler returned a single item and that item would be
a undefined reference, CommandStack will prevent it from being tracked
as dirty.

However, if a command handler returns an array containing a undefined
element, the array will simply be concated to the dirty collection
without filter the undefined element.

The above can be reproduced in the current version of bpmn-js (v0.14.1)

  1. Create a gateway
  2. Create a second element (task for example)
  3. Create a sequence flow connection from the gateway to the task
  4. Make the sequence flow default
  5. Remove the sequence flow connection
  6. Create a sequence flow connection from the gateway to the task (new)
  7. Make the sequence flow default

The 7th step will produce an exception as the UpdatePropertiesHandler
will return undefined as the first sequence flow no longer exists, but
the gateway's default property was not cleared in step 5.