feat(modeling): support append + attach
Be able to append and attach at the same time.

This allows us to support a richer modeling interaction where users
may create, attach and connect elements at the same time.

This changes the following <Modeling> APIs to move attach information
to the <hints> parameter (as it is truely optional):

* Modeling#moveElements
* Modeling#createShape
* Modeling#appendShape

Pass { attach: (true|false|... } via hint now instead of via a
boolean to the API.

Related to #242

BREAKING CHANGE:

* Modeling#{moveElements|createShape|appendShape} now receive
  { attach: true } via hints rather than isAttach parameter.
  Tool builders who invoke these APIs with attach support enabled
  __MUST__ switch to the new API (cf. changes carried out in this
  commit).