feat(create): support { attach, connect } during create
This adds support for both, attach and connect to be triggered
via Create.

On hover, Create will now query `connection.create` to figure
out whether or not the creation of a connection is possible
between the new shape and the source (if given).

During `connection.create` the target is not actually added yet,
instead the available attach / target meta data is passed to the
rule via hints = { targetParent, targetAttach }.

Related to #242

BREAKING CHANGE:

* Rule providers must rewrite their `connection.create` rules to
  take into account that the connection target may not be attached
  yet.