refactor(modeling): move labels + attachers with closure
Prior to this commit we adopted a complicated mechanism
to move attachers and labels after the actual move
closure.

This commit simplifies the matter: We now add attachers
and labels to the move closure and let our existing
move mechanisms handle their movements, too.

As a result, users can rely on the fact that all
elements moved once

* connection layouting is triggered
* the elements.move postExecute phase is triggered

:tada: