feat(grid-snapping): separate visuals and disable per default
Grid snapping visuals are still part of the library but won't be
shipped with the user.

They may be used by including grid-snapping/visuals as a feature
in test cases.

As the dots are a debug feature, this commit simplifies the
configuration and decouples the core GridSnapping from the visual
grid.

Related to bpmn-io/bpmn-js#1074