test(bendpoints): verifying that side effect do not occur
Adds testing for the issue presented in #965.

But I cannot recapture the report behaviour in the issue outside testing for the root svg element (which fails on develop without the fix from #976). But for element.out on shapes I was unable to trigger the same occurrences, perhaps due to the mock setup.
Even when I copy the exact workings from the  codepen :/